pyproject.toml
[tool.poetry]
name = "ubpkg"
version = "0.1.0"
description = ""
authors = ["alex <alex@pdp7.net>"]

[tool.poetry.scripts]
ubpkg = 'ubpkg.cli:run'

[tool.poetry.dependencies]
python = "^3.9"
pygithub = "^2.2.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
ipython = "<8.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
